sedcli-kmip: removed openssl submodule
It is impractical to keep openssl submodule added to sedcli project
for multiple reasons. For example openssl dependency is also required
by libkmip submodule. There is no good way to point libkmip to use
openssl submodule instead of searching for it in OS installation path.
This patch removes openssl submodule from sedcli and instead relies on
OS installed version of it.

Signed-off-by: Andrzej Jakowski <andrzej.jakowski@intel.com>