libsed: opal_parser: Put token back to free list when error occurs
In the fucntion opal_get_next_token, we are neglecting the token allocated
when there is an error parsing or the response payload has ended.
As a result we keep missing out a token while deallocating tokens and
putting them back to the list of free tokens -> number of available tokens
keeps reducing every time by 1.
This patch fixes the issue.

Signed-off-by: Revanth Rajashekar <revanth.rajashekar@intel.com>